hp -> float 형으로 변환, 네브메쉬 주석
이후 데미지 계산에서 소숫점이 필요함으로, 네브메쉬가 적 피격을 이상하게 만듦

+++ TpsTemplet/Assets/Scripts/UI/GamePlayUI.cs
@@ -42,7 +42,7 @@ public class GamePlayUI : MonoBehaviour
         ammoText.text = $"{currentAmmo} / {maxAmmo}";
     }
 
-    void UpdateHPUI(int currentHP, int maxHP)
+    void UpdateHPUI(float currentHP, float maxHP)
     {
         hpText.text = $"{currentHP} / {maxHP}";
     }
